# routers/api/docs/swagger.yaml
basePath: /
definitions:
  Rsp:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/proto.CodeType'
        description: msg code
      message:
        description: msg info
        type: string
      requestId:
        description: trans id
        type: string
      results:
        description: response body
    type: object
...
  proto.CodeType:
    enum:
    - 200
    - 201
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    - 1010
    - 1011
    - 1012
    - 1013
    - 1014
    - 1015
    - 1016
    - 1017
    - 1018
    - 1021
    - 1022
    - 1027
    - 1028
    - 1029
    - 1030
    - 1031
    - 1032
    - 1036
    - 1037
    - 1038
    - 1046
    - 1050
    - 1061
    - 1062
    type: integer
    x-enum-comments:
      CodeCopyIdError: 文件夹不能拷贝到自己
      CodeFailedToConnect: 建立连接失败
      CodeFailedToConnectDB: 连接DB失败
      CodeFailedToCreateBucket: 失败去创建桶
      CodeFailedToCreateFolder: 失败去创建文件夹
      CodeFailedToCreateMultipartTask: 失败去创建任务
      CodeFailedToCreateSymlink: 失败去创建符号链接
      CodeFailedToDeleteUser: 删除用户失败
      CodeFailedToGetUsedStorage: 获取用户空间使用量失败
      CodeFailedToInitUser: 失败去初始化用户
      CodeFailedToOpenFile: 打开文件失败
      CodeFailedToOperateDB: 失败去操作数据库
      CodeFailedToOptMinio: 失败去写文件
      CodeFailedToSaveFile: 失败去保存数据
      CodeFileExist: 文件已存在
      CodeFileNotExist: 文件不存在
      CodeFolderDepthTooLong: 文件夹层数超过5层
      CodeFolderExist: 文件夹已存在
      CodeFolderNotExist: 文件夹不存在
      CodeGetAsyncTaskInfoFailed: 获取异步任务状态失败
      CodeMultipartRangeUploaded: 分片范围已上传
      CodeMultipartTaskCompleteErr: 合并错误
      CodeMultipartTaskHashErr: 文件hash校验错误
      CodeMultipartTaskNotFound: 未找到任务
      CodeMultipartTaskOverlap: 任务重叠
      CodeMultipartTaskRangeErr: 上传范围错误
      CodeMultipartUploadingConflit: 分片范围上传冲突
      CodeNotEnoughSpace: 空间不够，不上传
      CodeParamErr: 参数错误
      CodeReqParamErr: 请求参数错误(推荐)
      CodeUserIdError: 用户Id的错误
      CodeWriteRedisFailed: 写入redis 失败
      CodeZipFileFailed: 压缩文件失败
    x-enum-varnames:
    - CodeOk
    - CodeCreateAsyncTaskSuccess
    - CodeParamErr
    - CodeReqParamErr
    - CodeFileNotExist
    - CodeFailedToOpenFile
    - CodeFolderNotExist
    - CodeFailedToConnect
    - CodeFailedToConnectDB
    - CodeFailedToOptMinio
    - CodeFailedToCreateBucket
    - CodeFailedToSaveFile
    - CodeFailedToOperateDB
    - CodeFailedToCreateFolder
    - CodeFileExist
    - CodeFolderExist
    - CodeFolderDepthTooLong
    - CodeUserIdError
    - CodeFailedToInitUser
    - CodeFailedToDeleteUser
    - CodeFailedToGetUsedStorage
    - CodeCopyIdError
    - CodeFailedToCreateMultipartTask
    - CodeMultipartTaskNotFound
    - CodeMultipartTaskOverlap
    - CodeMultipartTaskRangeErr
    - CodeMultipartTaskHashErr
    - CodeMultipartTaskCompleteErr
    - CodeNotEnoughSpace
    - CodeMultipartRangeUploaded
    - CodeMultipartUploadingConflit
    - CodeWriteRedisFailed
    - CodeZipFileFailed
    - CodeFailedToCreateSymlink
    - CodeGetAsyncTaskInfoFailed
  proto.CompleteMultipartTaskReq:
    properties:
      uploadId:
        type: string
    required:
    - uploadId
    type: object
  proto.CompleteMultipartTaskRsp:
    properties:
      betag:
        type: string
      bucketName:
        type: string
      category:
        type: string
      createdAt:
        type: integer
      executable:
        type: boolean
      ext:
        items:
          type: integer
        type: array
      fileCount:
        type: integer
      isDir:
        type: boolean
      mime:
        type: string
      modifyAt:
        type: integer
      name:
        type: string
      operationAt:
        type: integer
      parentUuid:
        type: string
      path:
        type: string
      size:
        type: integer
      tags:
        type: string
      transactionId:
        type: integer
      trashed:
        description: 0-normal; 1-Logical delete, put into the recycle bin; 2-Has been
          cleared from the recycle bin and is to be physically deleted
        type: integer
      userId:
        type: integer
      uuid:
        type: string
      version:
        type: integer
    type: object
...
  proto.CreateFolderReq:
    properties:
      currentDirUuid:
        description: current folder's uuid
        type: string
      folderName:
        description: folder name
        type: string
    type: object
  proto.CreateMultipartTaskConflictRsp:
    properties:
      albumId:
        type: integer
      betag:
        maxLength: 34
        minLength: 32
        type: string
      businessId:
        type: integer
      createTime:
        type: integer
      fileName:
        type: string
      folderId:
        type: string
      folderPath:
        type: string
      mime:
        type: string
      modifyTime:
        type: integer
      size:
        type: integer
      uploadId:
        type: string
      uploadedParts:
        items:
          $ref: '#/definitions/proto.Part'
        type: array
      uploadingParts:
        items:
          $ref: '#/definitions/proto.Part'
        type: array
    required:
    - betag
    - fileName
    - size
    type: object
  proto.CreateMultipartTaskReq:
    properties:
      albumId:
        type: integer
      betag:
        maxLength: 34
        minLength: 32
        type: string
      businessId:
        type: integer
      createTime:
        type: integer
      fileName:
        type: string
      folderId:
        type: string
      folderPath:
        type: string
      mime:
        type: string
      modifyTime:
        type: integer
      size:
        type: integer
    required:
    - betag
    - fileName
    - size
    type: object
  proto.CreateMultipartTaskRsp:
    properties:
      completeInfo:
        allOf:
        - $ref: '#/definitions/proto.FileInfo'
        description: task completed
      conflictInfo:
        allOf:
        - $ref: '#/definitions/proto.CreateMultipartTaskConflictRsp'
        description: task exists
      rspType:
        description: rsp type
        type: integer
      succInfo:
        allOf:
        - $ref: '#/definitions/proto.CreateMultipartTaskSuccRsp'
        description: task info
    type: object
  proto.CreateMultipartTaskSuccRsp:
    properties:
      partSize:
        type: integer
      uploadId:
        type: string
    type: object
...
  proto.DeleteMultipartTaskReq:
    properties:
      uploadId:
        type: string
    required:
    - uploadId
    type: object
...
  proto.FileInfo:
    properties:
      betag:
        type: string
      bucketName:
        type: string
      category:
        type: string
      createdAt:
        type: integer
      executable:
        type: boolean
      ext:
        items:
          type: integer
        type: array
      fileCount:
        type: integer
      isDir:
        type: boolean
      mime:
        type: string
      modifyAt:
        type: integer
      name:
        type: string
      operationAt:
        type: integer
      parentUuid:
        type: string
      path:
        type: string
      size:
        type: integer
      tags:
        type: string
      transactionId:
        type: integer
      trashed:
        description: 0-normal; 1-Logical delete, put into the recycle bin; 2-Has been
          cleared from the recycle bin and is to be physically deleted
        type: integer
      userId:
        type: integer
      uuid:
        type: string
      version:
        type: integer
    type: object
...
  proto.FolderInfo:
    properties:
      name:
        type: string
      operationAt:
        type: integer
      path:
        type: string
      size:
        type: integer
    type: object
...
  proto.ListMultipartReq:
    properties:
      uploadId:
        type: string
    required:
    - uploadId
    type: object
  proto.ListMultipartRsp:
    properties:
      uploadedParts:
        items:
          $ref: '#/definitions/proto.Part'
        type: array
      uploadingParts:
        items:
          $ref: '#/definitions/proto.Part'
        type: array
    type: object
...
  proto.Part:
    properties:
      end:
        type: integer
      start:
        minimum: 0
        type: integer
    required:
    - end
    type: object
...
info:
  contact:
    email: service@ao.space
    name: AO.space
    url: https://ao.space/
  description: This is AO.space aofs OpenAPI reference document.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: aofs apis
  version: "1.0"
paths:
...
  /space/v1/api/folder/create:
    post:
      consumes:
      - application/json
      description: Create Folder
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: params
        in: body
        name: createFolderReq
        required: true
        schema:
          $ref: '#/definitions/proto.CreateFolderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/proto.FileInfo'
              type: object
      summary: Create Folder
      tags:
      - Folder
  /space/v1/api/folder/info:
    get:
      consumes:
      - application/json
      description: Get folder details
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: folder's uuid
        in: query
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/proto.FolderInfo'
              type: object
      summary: Get folder details
      tags:
      - Folder
...
  /space/v1/api/multipart/complete:
    post:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: request id
        in: query
        name: requestId
        required: true
        type: string
      - description: params
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/proto.CompleteMultipartTaskReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/proto.CompleteMultipartTaskRsp'
              type: object
      summary: Complte multipart task
      tags:
      - Multipart
  /space/v1/api/multipart/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: request id
        in: query
        name: requestId
        required: true
        type: string
      - description: params
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/proto.CreateMultipartTaskReq'
      produces:
      - application/json
      responses:
        "200":
          description: task info
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/proto.CreateMultipartTaskRsp'
              type: object
      summary: Creating a multipart task
      tags:
      - Multipart
  /space/v1/api/multipart/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: request id
        in: query
        name: requestId
        required: true
        type: string
      - description: params
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/proto.DeleteMultipartTaskReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Rsp'
      summary: Delete multipart task
      tags:
      - Multipart
  /space/v1/api/multipart/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: request id
        in: query
        name: requestId
        required: true
        type: string
      - description: param
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/proto.ListMultipartReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/proto.ListMultipartRsp'
              type: object
      summary: Query multipart task info
      tags:
      - Multipart
  /space/v1/api/multipart/upload:
    post:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: request id
        in: query
        name: requestId
        required: true
        type: string
      - in: query
        name: end
        required: true
        type: integer
      - in: query
        name: md5sum
        required: true
        type: string
      - in: query
        minimum: 0
        name: start
        type: integer
      - in: query
        name: uploadId
        required: true
        type: string
      - description: part content
        in: body
        name: object
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Rsp'
      summary: Upload multipart data
      tags:
      - Multipart
...
swagger: "2.0"